Backend/requirements.txt
transformers
torch            
sqlalchemy
psycopg2-binary
Flask>=2.3
Flask-Mail>=0.9.1
Flask-JWT-Extended>=4.4.4
Flask-SQLAlchemy>=3.0
python-dotenv>=1.0
itsdangerous>=2.1
Werkzeug>=2.3
pytesseract Pillow